[DM-23942] Use get_redis_client everywhere
Whenever creating a new Redis client, use the get_redis_client
utility function, and move it into jwt_authorizer.util.  This will
make it easier for the test suite to override a single function
and use fakeredis.
