+++ src/jwt_authorizer/util.py
@@ -2,7 +2,14 @@
 
 from __future__ import annotations
 
-__all__ = ["add_padding"]
+from typing import TYPE_CHECKING
+
+import redis
+
+if TYPE_CHECKING:
+    from flask import Flask
+
+__all__ = ["add_padding", "get_redis_client"]
 
 
 def add_padding(encoded: str) -> str:
@@ -23,3 +30,21 @@ def add_padding(encoded: str) -> str:
         return encoded + ("=" * (4 - underflow))
     else:
         return encoded
+
+
+def get_redis_client(app: Flask) -> redis.Redis:
+    """Get a Redis client from the Flask application pool.
+
+    Exists primarily to be overridden by tests.
+
+    Parameters
+    ----------
+    app : `flask.Flask`
+        The Flask application.
+
+    Returns
+    -------
+    redis_client : `redis.Redis`
+        A Redis client.
+    """
+    return redis.Redis(connection_pool=app.redis_pool)
